fixing lookup context provider
