+++ src/components/Root.tsx
@@ -6,20 +6,39 @@ import { BrowserRouter as Router, Switch, Route } from 'react-router-dom'
 import shuffle from '../utils/FisherYatesShuffle'
 
 // core logic
-import fetch, { Recipe } from '../lib/data/fetch'
+import fetch from '../lib/data/fetch'
 import questions from '../lib/data/questions'
+import RecipeList, {
+  RecipeList as RecipeListType,
+} from '../lib/core/RecipeList'
 
 // other components
+import LookupContext, { LookupTables, RecipeLookup, RecipeByID } from '../utils/LookupContext'
 import Header from './header/Header'
 import Home from './home/Home'
 import RecipePage from './RecipePage'
 
 // import styles from './Root.module.sass'
 
-const Root = () => {
+const buildByUrlTable = (recipeList: RecipeListType): RecipeLookup => {
+  const byUrl: RecipeLookup = {}
+
+  Object.values(recipeList.allByID).forEach(
+    (recipe) => (byUrl[recipe.shortLink] = recipe.id)
+  )
+
+  return byUrl
+}
+
+export default () => {
   const [recipes, setRecipes] = React.useState({
-    data: [] as Array<Recipe>,
+    data: {} as RecipeListType,
     loaded: false,
+  })
+
+  const [ lookupTables, setLookupTables ] = React.useState<LookupTables>({
+    recipeByUrl: {} as RecipeLookup,
+    recipeByID: {} as RecipeByID
   })
 
   // calling setState later allows for component to load while
@@ -33,7 +52,12 @@ const Root = () => {
     fetch
       .recipes()
       .then((res) => {
-        const recipeList = shuffle(res.unwrap(), Math.random)
+        const recipeList = RecipeList.create(shuffle(res.unwrap(), Math.random))
+
+        setLookupTables({
+          recipeByUrl: buildByUrlTable(recipeList),
+          recipeByID: recipeList.allByID
+        })
 
         setRecipes({
           data: recipeList,
@@ -47,6 +71,7 @@ const Root = () => {
   }, [])
 
   return (
+    <LookupContext.Provider value={lookupTables}>
     <Router>
       <Header />
       <Switch>
@@ -63,7 +88,6 @@ const Root = () => {
         </Route>
       </Switch>
     </Router>
+    </LookupContext.Provider>
   )
 }
-
-export default Root
